Add files via upload
Finalized script. It is capable of automatically detecting and installing environment dependencies during the initial execution.

This file could be utilized by Platypus to create an integrated app. Jellyfin Media Player.app should be selected as "Bundled Files", and it is recommended to set "Interface" to "None" and check "Run in background" only.